modularise version control test
