+++ ui-tests/VersionControl/VersionControlComponents.test.tsx
@@ -2,7 +2,10 @@ import { mount, shallow } from "enzyme"
 import { shallowToJson } from "enzyme-to-json"
 import * as React from "react"
 
-import { Branch, BranchNameContainer } from "./../browser/src/UI/components/VersionControl/Branch"
+import {
+    Branch,
+    BranchNameContainer,
+} from "./../../browser/src/UI/components/VersionControl/Branch"
 
 describe("<Branch />", () => {
     const diff = {
